Moved service interfaces to own files and to core/interfaces folder
