+++ src/Chirp.Web/Pages/Follow.cshtml.cs
@@ -1,4 +1,5 @@
-﻿using Chirp.Infrastructure.Entities;
+﻿using Chirp.Core.Interfaces;
+using Chirp.Infrastructure.Entities;
 using Chirp.Web.Services;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
