+++ src/Chirp.Web/Pages/Public.cshtml.cs
@@ -1,4 +1,5 @@
 ﻿using Chirp.Core.DTO;
+using Chirp.Core.Interfaces;
 using Chirp.Infrastructure.Entities;
 using Chirp.Web.Pages.Shared;
 using Microsoft.AspNetCore.Mvc;
