+++ src/Chirp.Web/Pages/Shared/TimelineBaseModel.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using System.ComponentModel.DataAnnotations;
+using Chirp.Core.Interfaces;
 using Microsoft.AspNetCore.Authentication;
 
 namespace Chirp.Web.Pages.Shared;
