+++ src/Chirp.Web/Pages/UserTimeline.cshtml.cs
@@ -1,4 +1,5 @@
-﻿using Chirp.Infrastructure.Entities;
+﻿using Chirp.Core.Interfaces;
+using Chirp.Infrastructure.Entities;
 using Chirp.Web.Pages.Shared;
 using Microsoft.AspNetCore.Mvc;
 using Chirp.Web.Services;
